refactor : separate comment & reply feature
