+++ src/components/modal/post/ModalStyles.ts
